+++ src/models/redux.model.ts
@@ -1,0 +1,16 @@
+import { Dispatch, AnyAction } from 'redux'
+import { ActionCreatorDefinition } from '..'
+
+export interface StateToProps<TState = any> {
+  <T extends { [key: string]: any } = {}>(map: (store: TState) => T): (
+    store: TState,
+  ) => T
+}
+
+export interface DispatchToProps<
+  TDispatchAction extends AnyAction = AnyAction
+> {
+  <T extends { [key: string]: ActionCreatorDefinition }>(map: T): (
+    dispatch: Dispatch<TDispatchAction>,
+  ) => T
+}